Rm `thiserror` dep
Including `thiserror` in a lib forces the users of the lib to build `thiserror`
and it's dependencies. Namely `proc-macro2`, `quote` and `syn`.

These dependencies will more than likely be included through another dependecy path,
but given that there are only three Error types (one only for build.rs) and the
expansion of the Error definitions are simple, it's better to drop `thiserror`
in favor of keeping the dependecy set as small as possible.