fix(RFC3966): disallow whitespace per RFC3966
No whitespace character appears in the syntax described in RFC3966.

Previously, "+1 650-253-0000" was parsed, but panicked later when
validating the number. It parsed "+1 650" as the prefix, which is not a
valid prefix. Disallowing whitespace here prevents parsing the
phone number using the RFC3966 format. Instead, the regex based parser is
used.

Fixes #83